+++ RELEASES.md
@@ -9,6 +9,7 @@ The ongoing changes that are not released, are visible in the git commits and gi
 The TODO section is part of the [README.md](https://github.com/bestia-dev/rust_project_name).  
 
 ## Unreleased
+- auto_md_to_doc_comments include
 
 ## Version 2024.302.2056 (2024-03-02)